#234 Extract Sirius adaptation on helper
