+++ doc/org.eclipse.emf.diffmerge.doc/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,2 @@
 eclipse.preferences.version=1
-encoding//src/EMF-DiffMerge.html=UTF-8
 encoding/<project>=UTF-8
