+++ plugins/org.eclipse.emf.diffmerge.sirius/src/org/eclipse/emf/diffmerge/sirius/SiriusScope.java
@@ -18,7 +18,6 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.NullProgressMonitor;
@@ -33,17 +32,14 @@ import org.eclipse.emf.ecore.EPackage;
 import org.eclipse.emf.ecore.EReference;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.ResourceSet;
-import org.eclipse.emf.ecore.util.EcoreUtil;
 import org.eclipse.emf.edit.domain.EditingDomain;
 import org.eclipse.gmf.runtime.notation.NotationPackage;
 import org.eclipse.sirius.business.api.helper.SiriusUtil;
-import org.eclipse.sirius.business.api.image.RichTextAttributeRegistry;
 import org.eclipse.sirius.business.api.query.DRepresentationQuery;
 import org.eclipse.sirius.business.api.resource.ResourceDescriptor;
 import org.eclipse.sirius.business.api.session.Session;
 import org.eclipse.sirius.business.api.session.SessionManager;
 import org.eclipse.sirius.diagram.DiagramPackage;
-import org.eclipse.sirius.diagram.WorkspaceImage;
 import org.eclipse.sirius.viewpoint.DAnalysis;
 import org.eclipse.sirius.viewpoint.DRepresentation;
 import org.eclipse.sirius.viewpoint.DRepresentationDescriptor;
@@ -83,8 +79,8 @@ public class SiriusScope extends GMFScope {
   protected final Map<String, DRepresentationDescriptor> _idToDescriptor =
       new HashMap<String, DRepresentationDescriptor>();
   
-  /** @see org.eclipse.emf.diffmerge.sirius.SiriusScope getMainProjectName */
-  private HashMap<EObject, String> _mainProjectNames = new HashMap<>();
+  /** An helper for image helpers */
+  private SiriusImageHelper siriusImageHelper = new SiriusImageHelper();
   
   /**
    * Constructor
@@ -150,18 +146,7 @@ public class SiriusScope extends GMFScope {
   @Override
   public List<Object> get(EObject source_p, EAttribute attribute_p) {
     List<Object> objects = super.get(source_p, attribute_p);
-    
-    if (RichTextAttributeRegistry.INSTANCE.getEAttributes().contains(attribute_p)) {
-      objects = objects.stream().filter(String.class::isInstance).map((e) -> {
-        return replaceRichtextToRelative(source_p, (String) e);
-      }).collect(Collectors.toList());
-  
-    } else if (attribute_p == DiagramPackage.Literals.WORKSPACE_IMAGE__WORKSPACE_PATH) {
-      objects = objects.stream().filter(String.class::isInstance).map((e) -> {
-        return replaceWorkspaceImageToRelative(source_p, (String) e);
-      }).collect(Collectors.toList());
-    }
-
+    objects = siriusImageHelper.adaptGetValue(source_p, attribute_p, objects);
     return objects;
   }
 
@@ -171,149 +156,20 @@ public class SiriusScope extends GMFScope {
    */
   @Override
   public boolean add(EObject source_p, EAttribute attribute_p, Object value_p) {
-    if (value_p instanceof String) {
-      if (RichTextAttributeRegistry.INSTANCE.getEAttributes().contains(attribute_p)) {
-        String replacedValue = replaceRichtextToHardcodedProject(source_p, (String) value_p);
-        return super.add(source_p, attribute_p, replacedValue);
-  
-      } else if (attribute_p == DiagramPackage.Literals.WORKSPACE_IMAGE__WORKSPACE_PATH) {
-        String replacedValue = replaceWorkspaceImageToHardcodedProject(source_p, (String) value_p);
-        return super.add(source_p, attribute_p, replacedValue);
-      }
-    }
-    return super.add(source_p, attribute_p, value_p);
+    Object value = siriusImageHelper.adaptAddValue(source_p, attribute_p, value_p);
+    return super.add(source_p, attribute_p, value);
   }
-
+  
   /**
    * @see org.eclipse.emf.diffmerge.impl.scopes.AbstractEditableModelScope#remove(org.eclipse.emf.ecore.EObject,
    *      org.eclipse.emf.ecore.EAttribute, java.lang.Object)
    */
   @Override
   public boolean remove(EObject source_p, EAttribute attribute_p, Object value_p) {
-    if (value_p instanceof String) {
-      if (RichTextAttributeRegistry.INSTANCE.getEAttributes().contains(attribute_p)) {
-        String replacedValue = replaceRichtextToHardcodedProject(source_p, (String) value_p);
-        return super.remove(source_p, attribute_p, replacedValue);
-  
-      } else if (attribute_p == DiagramPackage.Literals.WORKSPACE_IMAGE__WORKSPACE_PATH) {
-        String replacedValue = replaceWorkspaceImageToHardcodedProject(source_p, (String) value_p);
-        return super.remove(source_p, attribute_p, replacedValue);
-      }
-    }
-    return super.remove(source_p, attribute_p, value_p);
+    Object value = siriusImageHelper.adaptRemoveValue(source_p, attribute_p, value_p);
+    return super.remove(source_p, attribute_p, value);
   }
 
-  /**
-   * Replace paths of images located in the root project by a relative path to avoid detecting differences on models
-   * with different names
-   * 
-   * @param source_p  the element
-   * @param value_p the richtext description that can contain image paths
-   * @implNote Notice the ':' we need to keep protocol in it (commit: for instance)
-   * @return a richtext description containing relative paths on images located in the root project
-   */
-  private String replaceRichtextToRelative(EObject source_p, String value_p) {
-    String projectName = getMainProjectName(source_p);
-    if (projectName == null) {
-      return value_p;
-    }
-    return value_p.replaceAll("src=\"([^/]+:/)?" + projectName + "/", "src=\"$1./"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-  }
-
-  /**
-   * Replace relative paths of images by an absolute path toward the root project of the source_p element
-   * 
-   * @param source_p the element
-   * @param value_p the richtext description that can contain image paths
-   * @return a richtext description containing relative paths on images located in the root project
-   */
-  private String replaceRichtextToHardcodedProject(EObject source_p, String value_p) {
-    String projectName = getMainProjectName(source_p);
-    if (projectName == null) {
-      return value_p;
-    }
-    return value_p.replaceAll("src=\"\\./", "src=\"" + projectName + "/"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-  }
-
-  /**
-   * Replace paths of {@link WorkspaceImage} located in the root project by a relative path to avoid detecting
-   * differences on models with different names
-   * 
-   * @param source_p
-   *          the element
-   * @param value_p
-   *          the {@link WorkspaceImage} workspace path
-   * @implNote Notice the ':' we need to keep protocol in it (commit: for instance)
-   * @return a workspace path containing relative path if image was located in the root project
-   */
-  private String replaceWorkspaceImageToRelative(EObject source_p, String value_p) {
-    String projectName = getMainProjectName(source_p);
-    if (projectName == null) {
-      return value_p;
-    }
-    return value_p.replaceAll("^([^/]+:/)?" + projectName + "/", "$1./"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-  }
-
-  /**
-   * Replace relative paths of {@link WorkspaceImage} by an absolute path toward the root project of the source_p
-   * element
-   * 
-   * @param source_p the element
-   * @param value_p the {@link WorkspaceImage} workspace path
-   * @return a workspace path containing absolute path if image was located in the root project
-   */
-  private String replaceWorkspaceImageToHardcodedProject(EObject source_p, String value_p) {
-    String projectName = getMainProjectName(source_p);
-    if (projectName == null) {
-      return value_p;
-    }
-    return value_p.replaceAll("^\\./", projectName + "/"); //$NON-NLS-1$ //$NON-NLS-2$
-  }
-
-  /**
-   * For a given object, retrieve its project
-   */
-  private String getProjectName(EObject source_p) {
-    EObject root = EcoreUtil.getRootContainer(source_p, true);
-    if (root != null) {
-      Resource resource = root.eResource();
-      if (resource != null) {
-        return getProjectFromUri(resource.getURI());
-      }
-    }
-    return null;
-  }
-
-  /**
-   * For a given resource URI, retrieve the root segment describing the root project name
-   * 
-   * @apiNote this api suppose that the given resource is directly located in it's root project. 
-   *          (which is the case in caller methods)
-   * @implNote URI can be from platform:/resource, commit:/ or other protocol such as cdo:/
-   */
-  protected String getProjectFromUri(URI uri_p) {
-    try {
-      return uri_p.segment(uri_p.segmentCount() - 2);
-    } catch (IndexOutOfBoundsException e) {
-      return null;
-    }
-  }
-  
-  /**
-   * For an element, retrieve it's project name
-   * 
-   * @implSpec Note that the project name containing the element source_p might be different than the IProject containing the
-   *           Sirius session. Here we lookup for the project name owning the resource containing the element
-   */
-  protected String getMainProjectName(EObject source_p) {
-    EObject root = EcoreUtil.getRootContainer(source_p);
-    if (!_mainProjectNames.containsKey(root)) {
-      String string = getProjectName(root);
-      _mainProjectNames.put(root, string);
-    }
-    return _mainProjectNames.get(root);
-  }
-  
   /**
    * @see org.eclipse.emf.diffmerge.impl.scopes.AbstractModelScope#getContainer(org.eclipse.emf.ecore.EObject)
    */
